gw_spaceheat/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
appnope==0.1.3
    # via ipython
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
black==22.3.0
    # via ipython
click==8.1.3
    # via black
decorator==5.1.1
    # via ipython
executing==0.8.3
    # via stack-data
ipython==8.0.0
    # via -r requirements/dev.in
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.3
    # via ipython
mypy-extensions==0.4.3
    # via black
paho-mqtt==1.6.1
    # via -r requirements/dev.in
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pendulum==2.1.2
    # via -r requirements/dev.in
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pika==1.1.0
    # via -r requirements/dev.in
platformdirs==2.5.2
    # via black
prompt-toolkit==3.0.29
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.12.0
    # via ipython
python-dateutil==2.8.2
    # via pendulum
pytz==2022.1
    # via -r requirements/dev.in
pytzdata==2020.1
    # via pendulum
six==1.16.0
    # via python-dateutil
stack-data==0.2.0
    # via ipython
tomli==2.0.1
    # via black
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.2.0
    # via black
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
